backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
starlette>=0.36.3
typing-extensions>=4.9.0
openai>=1.3.0
tiktoken>=0.9.0
azure-cosmos>=4.5.1
sqlalchemy>=2.0.25
alembic>=1.13.1
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
cryptography>=42.0.0
twilio>=8.10.0
email-validator>=2.1.0
reportlab>=4.0.8
aiofiles>=23.2.1
pillow>=11.2.1
requests>=2.31.0
httpx>=0.26.0
python-dateutil>=2.8.2
pytz>=2023.3
arrow>=1.3.0
orjson>=3.9.12
pandas>=2.1.4
numpy>=1.24.3
scipy>=1.11.4
scikit-learn>=1.3.2
matplotlib>=3.8.2
seaborn>=0.13.0
gunicorn>=21.2.0
prometheus-client>=0.19.0
python-json-logger>=2.0.7
redis>=5.0.1
structlog>=23.2.0
loguru>=0.7.2
cachetools>=5.3.2
asyncio-throttle>=1.0.2
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
pytest-cov>=4.1.0
coverage>=7.4.0
mock>=5.1.0
unittest-xml-reporting>=3.2.0
pylint>=3.0.0
autopep8>=2.0.0
black>=23.12.0
isort>=5.13.0
flake8>=7.0.0 